+++ docs/vocs.config.ts
@@ -25,6 +25,9 @@ export default defineConfig({
         { text: "Balance", link: "/components/Balance" },
         { text: "AddressInput", link: "/components/AddressInput" },
         { text: "EtherInput", link: "/components/EtherInput" },
+        { text: "BaseInput", link: "/components/BaseInput" },
+        { text: "Styling", link: "/components/Styling" },
+        { text: "Theming", link: "/components/Theming" },
       ],
     },
     {